Refactor: Header 컴포넌트에서 검색 Modal 분리 Add: Redux Persist 라이브러리 추가 및 설정 추가
