+++ package.json
@@ -24,6 +24,7 @@
     "react-responsive": "^9.0.2",
     "react-router-dom": "^6.6.1",
     "react-scripts": "5.0.1",
+    "redux-persist": "^6.0.0",
     "styled-components": "^5.3.10",
     "styled-react-modal": "^3.0.1",
     "styled-reset": "^4.4.5",
